fix(test): move mock session logic to setupTests in helpers.ts
Also test response of update user route
