+++ app/tests/api/user.test.ts
@@ -1,15 +1,12 @@
 import { PATCH as updateUser } from "@/app/api/v0/user/route";
 import { db } from "@/models";
 import { UserAttributes } from "@/models/User";
-import env from "@next/env";
 import assert from "node:assert";
-import { randomUUID } from "node:crypto";
-import { after, before, describe, it, mock } from "node:test";
-import { mockRequest } from "../helpers";
-import { AppSession, Auth } from "@/lib/auth";
+import { after, before, describe, it } from "node:test";
+import { mockRequest, setupTests, testUserID } from "../helpers";
 
 const userData: UserAttributes = {
-  userId: randomUUID(),
+  userId: testUserID,
   name: "TEST_USER_USER",
 };
 
@@ -20,26 +17,10 @@ const userUpdate = {
 
 describe("User API", () => {
   before(async () => {
-    const projectDir = process.cwd();
-    env.loadEnvConfig(projectDir);
+    setupTests();
     await db.initialize();
-    await db.models.User.destroy({ where: { name: userData.name } });
+    await db.models.User.destroy({ where: { userId: userData.userId } });
     await db.models.User.create(userData);
-
-    mock.method(Auth, "getServerSession", (): AppSession => {
-      const expires = new Date();
-      expires.setDate(expires.getDate() + 1);
-      return {
-        user: {
-          id: randomUUID(),
-          name: "Test User",
-          email: "test@example.com",
-          image: null,
-          role: "user",
-        },
-        expires: expires.toISOString(),
-      };
-    });
   });
 
   after(async () => {
@@ -50,6 +31,8 @@ describe("User API", () => {
     const req = mockRequest(userUpdate);
     const res = await updateUser(req, { params: {} });
     assert.equal(res.status, 200);
+    const data = await res.json();
+    assert.ok(data.success);
     const user = await db.models.User.findOne({
       where: { userId: userData.userId },
     });
